pyproject.toml
[project]
name = "ai-music-coproducer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.8.0",
    "python-reapy>=0.10.0",
]

[tool.uv.workspace]
members = [
    "mcp-server-demo",
]
